Editor: implement actions to transform properties of a single node via GuiObjectProperties

+++ editor/src/main/kotlin/hunternif/voxarch/editor/util/Types.kt
@@ -11,7 +11,7 @@ fun IntVec3.toVector3f() = Vector3f(x.toFloat(), y.toFloat(), z.toFloat())
 
 fun Vector3i.toVec3() = Vec3(x, y, z)
 fun Vector3f.toVec3() = Vec3(x, y, z)
-fun Vec3.toVector3i() = Vector3f(x.toFloat(), y.toFloat(), z.toFloat())
+fun Vec3.toVector3i() = Vector3i(x.toInt(), y.toInt(), z.toInt())
 fun Vec3.toVector3f() = Vector3f(x.toFloat(), y.toFloat(), z.toFloat())
 
 fun Vector3f.set(v: Vec3): Vector3f = set(v.x, v.y, v.z)
